+++ packages/in_app_purchase/in_app_purchase_android/android/src/main/java/io/flutter/plugins/inapppurchase/BillingClientFactory.java
@@ -17,10 +17,7 @@ interface BillingClientFactory {
    *
    * @param context The context used to create the {@link BillingClient}.
    * @param channel The method channel used to create the {@link BillingClient}.
-   * @param enablePendingPurchases Whether to enable pending purchases. Throws an exception if it is
-   *     false.
    * @return The {@link BillingClient} object that is created.
    */
-  BillingClient createBillingClient(
-      @NonNull Context context, @NonNull MethodChannel channel, boolean enablePendingPurchases);
+  BillingClient createBillingClient(@NonNull Context context, @NonNull MethodChannel channel);
 }