+++ packages/webview_flutter/webview_flutter_android/android/src/test/java/io/flutter/plugins/webviewflutter/WebViewTest.java
@@ -157,6 +157,52 @@ public class WebViewTest {
   }
 
   @Test
+  public void loadData() {
+    testHostApiImpl.loadData(
+        0L, "VGhpcyBkYXRhIGlzIGJhc2U2NCBlbmNvZGVkLg==", "text/plain", "base64");
+    verify(mockWebView)
+        .loadData("VGhpcyBkYXRhIGlzIGJhc2U2NCBlbmNvZGVkLg==", "text/plain", "base64");
+  }
+
+  @Test
+  public void loadDataWithNullValues() {
+    testHostApiImpl.loadData(
+        0L, "VGhpcyBkYXRhIGlzIGJhc2U2NCBlbmNvZGVkLg==", "<null-value>", "<null-value>");
+    verify(mockWebView).loadData("VGhpcyBkYXRhIGlzIGJhc2U2NCBlbmNvZGVkLg==", null, null);
+  }
+
+  @Test
+  public void loadDataWithBaseUrl() {
+    testHostApiImpl.loadDataWithBaseUrl(
+        0L,
+        "https://flutter.dev",
+        "VGhpcyBkYXRhIGlzIGJhc2U2NCBlbmNvZGVkLg==",
+        "text/plain",
+        "base64",
+        "about:blank");
+    verify(mockWebView)
+        .loadDataWithBaseURL(
+            "https://flutter.dev",
+            "VGhpcyBkYXRhIGlzIGJhc2U2NCBlbmNvZGVkLg==",
+            "text/plain",
+            "base64",
+            "about:blank");
+  }
+
+  @Test
+  public void loadDataWithBaseUrlAndNullValues() {
+    testHostApiImpl.loadDataWithBaseUrl(
+        0L,
+        "<null-value>",
+        "VGhpcyBkYXRhIGlzIGJhc2U2NCBlbmNvZGVkLg==",
+        "<null-value>",
+        "<null-value>",
+        "<null-value>");
+    verify(mockWebView)
+        .loadDataWithBaseURL(null, "VGhpcyBkYXRhIGlzIGJhc2U2NCBlbmNvZGVkLg==", null, null, null);
+  }
+
+  @Test
   public void loadUrl() {
     testHostApiImpl.loadUrl(0L, "https://www.google.com", new HashMap<>());
     verify(mockWebView).loadUrl("https://www.google.com", new HashMap<>());