+++ packages/webview_flutter/webview_flutter_platform_interface/lib/src/method_channel/webview_method_channel.dart
@@ -35,32 +35,34 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
   Future<bool?> _onMethodCall(MethodCall call) async {
     switch (call.method) {
       case 'javascriptChannelMessage':
-        final String channel = call.arguments['channel']!;
-        final String message = call.arguments['message']!;
+        final String channel = call.arguments['channel']! as String;
+        final String message = call.arguments['message']! as String;
         _javascriptChannelRegistry.onJavascriptChannelMessage(channel, message);
         return true;
       case 'navigationRequest':
         return await _platformCallbacksHandler.onNavigationRequest(
-          url: call.arguments['url']!,
-          isForMainFrame: call.arguments['isForMainFrame']!,
+          url: call.arguments['url']! as String,
+          isForMainFrame: call.arguments['isForMainFrame']! as bool,
         );
       case 'onPageFinished':
-        _platformCallbacksHandler.onPageFinished(call.arguments['url']!);
+        _platformCallbacksHandler
+            .onPageFinished(call.arguments['url']! as String);
         return null;
       case 'onProgress':
-        _platformCallbacksHandler.onProgress(call.arguments['progress']);
+        _platformCallbacksHandler.onProgress(call.arguments['progress'] as int);
         return null;
       case 'onPageStarted':
-        _platformCallbacksHandler.onPageStarted(call.arguments['url']!);
+        _platformCallbacksHandler
+            .onPageStarted(call.arguments['url']! as String);
         return null;
       case 'onWebResourceError':
         _platformCallbacksHandler.onWebResourceError(
           WebResourceError(
-            errorCode: call.arguments['errorCode']!,
-            description: call.arguments['description']!,
+            errorCode: call.arguments['errorCode']! as int,
+            description: call.arguments['description']! as String,
             // iOS doesn't support `failingUrl`.
-            failingUrl: call.arguments['failingUrl'],
-            domain: call.arguments['domain'],
+            failingUrl: call.arguments['failingUrl'] as String?,
+            domain: call.arguments['domain'] as String?,
             errorType: call.arguments['errorType'] == null
                 ? null
                 : WebResourceErrorType.values.firstWhere(
@@ -122,23 +124,24 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
 
   @override
   Future<bool> canGoBack() =>
-      _channel.invokeMethod<bool>("canGoBack").then((result) => result!);
+      _channel.invokeMethod<bool>('canGoBack').then((bool? result) => result!);
 
   @override
-  Future<bool> canGoForward() =>
-      _channel.invokeMethod<bool>("canGoForward").then((result) => result!);
+  Future<bool> canGoForward() => _channel
+      .invokeMethod<bool>('canGoForward')
+      .then((bool? result) => result!);
 
   @override
-  Future<void> goBack() => _channel.invokeMethod<void>("goBack");
+  Future<void> goBack() => _channel.invokeMethod<void>('goBack');
 
   @override
-  Future<void> goForward() => _channel.invokeMethod<void>("goForward");
+  Future<void> goForward() => _channel.invokeMethod<void>('goForward');
 
   @override
-  Future<void> reload() => _channel.invokeMethod<void>("reload");
+  Future<void> reload() => _channel.invokeMethod<void>('reload');
 
   @override
-  Future<void> clearCache() => _channel.invokeMethod<void>("clearCache");
+  Future<void> clearCache() => _channel.invokeMethod<void>('clearCache');
 
   @override
   Future<void> updateSettings(WebSettings settings) async {
@@ -152,7 +155,7 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
   Future<String> evaluateJavascript(String javascript) {
     return _channel
         .invokeMethod<String>('evaluateJavascript', javascript)
-        .then((result) => result!);
+        .then((String? result) => result!);
   }
 
   @override
@@ -164,7 +167,7 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
   Future<String> runJavascriptReturningResult(String javascript) {
     return _channel
         .invokeMethod<String>('runJavascriptReturningResult', javascript)
-        .then((result) => result!);
+        .then((String? result) => result!);
   }
 
   @override
@@ -180,7 +183,7 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
   }
 
   @override
-  Future<String?> getTitle() => _channel.invokeMethod<String>("getTitle");
+  Future<String?> getTitle() => _channel.invokeMethod<String>('getTitle');
 
   @override
   Future<void> scrollTo(int x, int y) {
@@ -200,17 +203,17 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
 
   @override
   Future<int> getScrollX() =>
-      _channel.invokeMethod<int>("getScrollX").then((result) => result!);
+      _channel.invokeMethod<int>('getScrollX').then((int? result) => result!);
 
   @override
   Future<int> getScrollY() =>
-      _channel.invokeMethod<int>("getScrollY").then((result) => result!);
+      _channel.invokeMethod<int>('getScrollY').then((int? result) => result!);
 
   /// Method channel implementation for [WebViewPlatform.clearCookies].
   static Future<bool> clearCookies() {
     return _cookieManagerChannel
         .invokeMethod<bool>('clearCookies')
-        .then<bool>((dynamic result) => result!);
+        .then<bool>((dynamic result) => result! as bool);
   }
 
   static Map<String, dynamic> _webSettingsToMap(WebSettings? settings) {