+++ packages/webview_flutter/webview_flutter_platform_interface/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 * Add an option to set the background color of the webview.
 
+## 1.5.2
+
+* Mirgrates from analysis_options_legacy.yaml to the more strict analysis_options.yaml.
+
 ## 1.5.1
 
 * Reverts the addition of `onUrlChanged`, which was unintentionally a breaking