+++ packages/webview_flutter/webview_flutter_wkwebview/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Adds an option to set the background color of the webview.
 * Migrates from `analysis_options_legacy.yaml` to `analysis_options.yaml`.
 * Integration test fixes.
+* Updates to webview_flutter_platform_interface version 1.5.2.
 
 ## 2.4.0